Basic watchdog timer setup
This is a part of the work to implement reliable OTA firmware
updates on the Tau V3s board. We start a 10-second watchdog count
down in U-Boot before passing control to the kernel. Once we reach
userspace, the kernel module sunxi_wdt.ko is immediately loaded
to re-initialize the watchdog. This allows the system to be
restarted if the kernel gets stuck.